Rename same_channel to identical_to
Renaming the test file in the process.
